+++ src/cvode/fcmix/fcvband.c
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.5 $
- * $Date: 2007-04-30 19:28:59 $
+ * $Revision: 1.6 $
+ * $Date: 2010-12-01 22:27:37 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Alan C. Hindmarsh and Radu Serban @ LLNL
  * -----------------------------------------------------------------
@@ -30,7 +30,7 @@
 #ifdef __cplusplus  /* wrapper to enable C++ usage */
 extern "C" {
 #endif
-  extern void FCV_BJAC(int*, int*, int*, int*,           /* N, MU, ML, EBAND */
+  extern void FCV_BJAC(long int*, long int*, long int*, long int*,  /* N,MU,ML,EBAND */
                        realtype*, realtype*, realtype*,  /* T, Y, FY         */
                        realtype*,                        /* BJAC             */
                        realtype*,                        /* H                */
@@ -66,16 +66,15 @@ void FCV_BANDSETJAC(int *flag, int *ier)
    passed as the column dimension of the corresponding array.
    Auxiliary data is assumed to be communicated by Common. */
 
-int FCVBandJac(int N, int mupper, int mlower,
-               realtype t,
-               N_Vector y, N_Vector fy, 
+int FCVBandJac(long int N, long int mupper, long int mlower,
+               realtype t, N_Vector y, N_Vector fy, 
                DlsMat J, void *user_data,
                N_Vector vtemp1, N_Vector vtemp2, N_Vector vtemp3)
 {
   int ier;
   realtype *ydata, *fydata, *jacdata, *v1data, *v2data, *v3data;
   realtype h;
-  int eband;
+  long int eband;
   FCVUserData CV_userdata;
 
   CVodeGetLastStep(CV_cvodemem, &h);
