+++ src/cvodes/cvodes_direct.c
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.9 $
- * $Date: 2008-04-18 19:42:40 $
+ * $Revision: 1.10 $
+ * $Date: 2010-12-01 22:30:43 $
  * ----------------------------------------------------------------- 
  * Programmer: Radu Serban @ LLNL
  * -----------------------------------------------------------------
@@ -46,12 +46,12 @@
  * =================================================================
  */
 
-static int cvDlsDenseJacBWrapper(int nB, realtype t,
+static int cvDlsDenseJacBWrapper(long int nB, realtype t,
                                  N_Vector yB, N_Vector fyB, 
                                  DlsMat JB, void *cvode_mem,
                                  N_Vector tmp1B, N_Vector tmp2B, N_Vector tmp3B);
 
-static int cvDlsBandJacBWrapper(int nB, int mupperB, int mlowerB, 
+static int cvDlsBandJacBWrapper(long int nB, long int mupperB, long int mlowerB, 
                                 realtype t, N_Vector yB, N_Vector fyB, 
                                 DlsMat Jac, void *cvode_mem, 
                                 N_Vector tmp1B, N_Vector tmp2B, N_Vector tmp3B);
@@ -250,7 +250,7 @@ int CVDlsGetNumRhsEvals(void *cvode_mem, long int *nfevalsLS)
  * CVDlsGetReturnFlagName returns the name associated with a CVDLS
  * return value.
  */
-char *CVDlsGetReturnFlagName(int flag)
+char *CVDlsGetReturnFlagName(long int flag)
 {
   char *name;
 
@@ -288,7 +288,7 @@ char *CVDlsGetReturnFlagName(int flag)
 /*
  * CVDlsGetLastFlag returns the last flag set in a CVDLS function.
  */
-int CVDlsGetLastFlag(void *cvode_mem, int *flag)
+int CVDlsGetLastFlag(void *cvode_mem, long int *flag)
 {
   CVodeMem cv_mem;
   CVDlsMem cvdls_mem;
@@ -332,7 +332,7 @@ int CVDlsGetLastFlag(void *cvode_mem, int *flag)
  * -----------------------------------------------------------------
  */ 
 
-int cvDlsDenseDQJac(int N, realtype t,
+int cvDlsDenseDQJac(long int N, realtype t,
                     N_Vector y, N_Vector fy, 
                     DlsMat Jac, void *data,
                     N_Vector tmp1, N_Vector tmp2, N_Vector tmp3)
@@ -340,7 +340,7 @@ int cvDlsDenseDQJac(int N, realtype t,
   realtype fnorm, minInc, inc, inc_inv, yjsaved, srur;
   realtype *tmp2_data, *y_data, *ewt_data;
   N_Vector ftemp, jthCol;
-  int j;
+  long int j;
   int retval = 0;
 
   CVodeMem cv_mem;
@@ -408,7 +408,7 @@ int cvDlsDenseDQJac(int N, realtype t,
  * -----------------------------------------------------------------
  */
 
-int cvDlsBandDQJac(int N, int mupper, int mlower,
+int cvDlsBandDQJac(long int N, long int mupper, long int mlower,
                    realtype t, N_Vector y, N_Vector fy, 
                    DlsMat Jac, void *data,
                    N_Vector tmp1, N_Vector tmp2, N_Vector tmp3)
@@ -416,7 +416,7 @@ int cvDlsBandDQJac(int N, int mupper, int mlower,
   N_Vector ftemp, ytemp;
   realtype fnorm, minInc, inc, inc_inv, srur;
   realtype *col_j, *ewt_data, *fy_data, *ftemp_data, *y_data, *ytemp_data;
-  int group, i, j, width, ngroups, i1, i2;
+  long int group, i, j, width, ngroups, i1, i2;
   int retval = 0;
 
   CVodeMem cv_mem;
@@ -633,7 +633,7 @@ int CVDlsSetBandJacFnB(void *cvode_mem, int which, CVDlsBandJacFnB jacB)
  */
 
 
-static int cvDlsDenseJacBWrapper(int nB, realtype t,
+static int cvDlsDenseJacBWrapper(long int nB, realtype t,
                                  N_Vector yB, N_Vector fyB, 
                                  DlsMat JB, void *cvode_mem,
                                  N_Vector tmp1B, N_Vector tmp2B, N_Vector tmp3B)
@@ -676,7 +676,7 @@ static int cvDlsDenseJacBWrapper(int nB, realtype t,
  * NOTE: data here contains cvode_mem
  */
 
-static int cvDlsBandJacBWrapper(int nB, int mupperB, int mlowerB, 
+static int cvDlsBandJacBWrapper(long int nB, long int mupperB, long int mlowerB, 
                                 realtype t, N_Vector yB, N_Vector fyB, 
                                 DlsMat JB, void *cvode_mem, 
                                 N_Vector tmp1B, N_Vector tmp2B, N_Vector tmp3B)
