+++ src/ida/fcmix/fidadense.c
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.5 $
- * $Date: 2007-04-30 19:29:00 $
+ * $Revision: 1.6 $
+ * $Date: 2010-12-01 22:37:20 $
  * ----------------------------------------------------------------- 
  * Programmer(s): Aaron Collier @ LLNL
  * -----------------------------------------------------------------
@@ -29,7 +29,7 @@
 extern "C" {
 #endif
 
-  extern void FIDA_DJAC(int*, 
+  extern void FIDA_DJAC(long int*, 
                         realtype*, realtype*, realtype*, realtype*,
                         realtype*, 
                         realtype*, realtype*, realtype*,
@@ -69,7 +69,7 @@ void FIDA_DENSESETJAC(int *flag, int *ier)
 
 /*************************************************/
 
-int FIDADenseJac(int N, realtype t, realtype c_j, 
+int FIDADenseJac(long int N, realtype t, realtype c_j, 
 		 N_Vector yy, N_Vector yp, N_Vector rr,
                  DlsMat Jac, void *user_data,
 		 N_Vector vtemp1, N_Vector vtemp2, N_Vector vtemp3)
